requirements.txt
google-generativeai>=0.3.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
